+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Fixes
 - [Issue #20] publish.sh runs the build twice
+- [Issue #16] Build throws unfriendly exception if ts-package-config.ts is missing
 
 ## [1.0.3]